a few mallocs less

+++ lib/V8/v8-utils.cpp
@@ -94,9 +94,9 @@ TRI_Utf8ValueNFC::TRI_Utf8ValueNFC (TRI_memory_zone_t* memoryZone, v8::Handle<v8
   _str(0), _length(0), _memoryZone(memoryZone) {
 
    v8::String::Value str(obj);
-   size_t str_len = str.length();
+   int str_len = str.length();
 
-   _str = TRI_normalize_utf16_to_NFC(_memoryZone, *str, str_len, &_length);
+   _str = TRI_normalize_utf16_to_NFC(_memoryZone, *str, (size_t) str_len, &_length);
 }
 
 TRI_Utf8ValueNFC::~TRI_Utf8ValueNFC () {
